moved metadata parsing methods to TikaUtils class. use of proper types for Page-Count (int) and X-OCR-Applied (boolean) instead of strings as by default for metadata keys
